chore(utils): Move `BigInt`-related functions to `bigint` module
